requirements.txt
langchain==0.0.228
openai==0.27.8
streamlit==1.24.0
duckduckgo-search==3.8.3
pypdf==3.12.0
sentence-transformers==2.2.2
docarray==0.32.1
python-dotenv